document_comparison/document_comparison_backend/requirements.txt
chromadb==0.3.29
docx==0.2.4
Flask==3.0.0
Flask_Bcrypt==1.0.1
flask_cors==4.0.0
flask_sqlalchemy==3.1.1
langchain==0.0.325
pandas==2.1.2
pypdf==3.17.0
openai==0.28.1
tiktoken
PyPDF2==3.0.1
python-dotenv==1.0.0
python_docx==0.8.11
tqdm==4.66.1
Werkzeug==3.0.1
matplotlib
openpyxl
tabulate
langchain_experimental